+++ index.js
@@ -40,33 +40,19 @@ for (const folder of commandFolders) {
 	}
 }
 
-// Événement déclenché lorsque le client est prêt
-client.once(Events.ClientReady, () => {
-	console.log(`Prêt ! Connecté en tant que ${client.user.tag}`);
-});
+// Récupération des événements
+const eventsPath = path.join(__dirname, "events");
+const eventFiles = fs.readdirSync(eventsPath).filter((file) => file.endsWith(".js"));
 
-// Événement déclenché lorsqu'un utilisateur envoie un message
-client.on(Events.InteractionCreate, async (interaction) => {
-	if (!interaction.isChatInputCommand()) return;
-
-	const command = interaction.client.commands.get(interaction.commandName);
-
-	if (!command) {
-		console.error(`No command matching ${interaction.commandName} was found.`);
-		return;
+for (const file of eventFiles) {
+	const filePath = path.join(eventsPath, file);
+	const event = require(filePath);
+	if (event.once) {
+		client.once(event.name, (...args) => event.execute(...args));
+	} else {
+		client.on(event.name, (...args) => event.execute(...args));
 	}
-
-	try {
-		await command.execute(interaction);
-	} catch (error) {
-		console.error(error);
-		if (interaction.replied || interaction.deferred) {
-			await interaction.followUp({ content: "There was an error while executing this command!", flags: MessageFlags.Ephemeral });
-		} else {
-			await interaction.reply({ content: "There was an error while executing this command!", flags: MessageFlags.Ephemeral });
-		}
-	}
-});
+}
 
 // Connexion du client à Discord
 client.login(token).catch((error) => {
